Simplify the filter logic into one function Problem: 4 filter functions is not KISS Solution: Combine filter functions into one function that accepts and returns all types of inputs and search types.
